requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
python-dotenv
email-validator
